requirements.txt
pandas
scikit-learn
psycopg2-binary
SQLAlchemy
python-dotenv
fastapi
boto3